Refactor project to use f32 for floating-point types
- Standardize floating-point precision across the project by changing all `f64` types to `f32`
- Remove redundant code and unify naming conventions for better readability and consistency
- Update image generation logic and dependent calculations to accommodate new precision standards
- Revise documentation and binary resources to align with codebase updates

+++ src/entities/sphere.zig
@@ -10,12 +10,12 @@ const HitRecord = hitable.HitRecord;
 
 pub const Sphere = struct {
     center: Vec3,
-    radius: f64,
+    radius: f32,
     mat: *Material,
 
     const Self = @This();
 
-    pub fn init(center: Vec3, radius: f64, material: *Material) Self {
+    pub fn init(center: Vec3, radius: f32, material: *Material) Self {
         return .{ .center = center, .radius = radius, .mat = material };
     }
